+++ game_patch/misc/misc.h
@@ -16,3 +16,4 @@ bool af_rfl_version(int version);
 bool rfl_version_minimum(int check_version);
 void evaluate_restrict_disable_ss();
 void evaluate_restrict_disable_muzzle_flash();
+void initialize_achievement_manager();