+++ game_patch/object/trigger.cpp
@@ -3,6 +3,7 @@
 #include <common/rfproto.h>
 #include <xlog/xlog.h>
 #include "../misc/misc.h"
+#include "../misc/achievements.h"
 #include "../rf/player/player.h"
 #include "../rf/level.h"
 #include "../rf/trigger.h"
@@ -39,6 +40,11 @@ FunHook<void(int, int)> send_trigger_activate_packet_to_all_players_hook{
 FunHook<void(rf::Trigger*, int32_t, bool)> trigger_activate_hook{
     0x004C0220,
     [](rf::Trigger* trigger, int32_t h_entity, bool skip_movers) {
+        // Check if this trigger activation should unlock an achievement
+        if (is_achievement_system_initialized()) {
+            achievement_check_trigger(trigger);
+        }
+
         // Check team
         rf::Player* player = rf::player_from_entity_handle(h_entity);
         const char* trigger_name = trigger->name.c_str();