+++ docs/CHANGELOG.md
@@ -62,6 +62,7 @@ Version 1.1.0 (Tamarack): Not yet released
 - Add location pinging for multiplayer team modes with new action bind
 - Add `+Location Pinging` setting to `$Clients Require Alpine Faction`
 - Add `cl_showping` command to display current ping when connected to a server
+- Add achievements framework
 
 
 Version 1.0.0 (Maple): Released Jan-25-2025